Feat | Huge refactor + filtering after appset generation
