+++ cmd/main.go
@@ -9,6 +9,7 @@ import (
 	"github.com/dag-andersen/argocd-diff-preview/pkg/argoapplication"
 	"github.com/dag-andersen/argocd-diff-preview/pkg/argocd"
 	"github.com/dag-andersen/argocd-diff-preview/pkg/diff"
+	"github.com/dag-andersen/argocd-diff-preview/pkg/duplicates"
 	"github.com/dag-andersen/argocd-diff-preview/pkg/extract"
 	"github.com/dag-andersen/argocd-diff-preview/pkg/git"
 	"github.com/dag-andersen/argocd-diff-preview/pkg/utils"
@@ -56,22 +57,28 @@ func run(opts *Options) error {
 	baseBranch := git.NewBranch(opts.BaseBranch, git.Base)
 	targetBranch := git.NewBranch(opts.TargetBranch, git.Target)
 
+	filterOptions := argoapplication.FilterOptions{
+		Selector:                  selectors,
+		FilesChanged:              filesChanged,
+		IgnoreInvalidWatchPattern: opts.IgnoreInvalidWatchPattern,
+	}
+
 	// Get applications for both branches
 	baseApps, targetApps, err := argoapplication.GetApplicationsForBranches(
 		opts.ArgocdNamespace,
 		baseBranch,
 		targetBranch,
 		fileRegex,
-		selectors,
-		filesChanged,
+		filterOptions,
 		opts.Repo,
-		opts.IgnoreInvalidWatchPattern,
 		redirectRevisions,
 	)
 	if err != nil {
 		log.Error().Msgf("❌ Failed to get applications")
 		return err
 	}
+
+	baseApps, targetApps = duplicates.RemoveDuplicates(baseApps, targetApps)
 
 	foundBaseApps := len(baseApps) > 0
 	foundTargetApps := len(targetApps) > 0
@@ -139,11 +146,19 @@ func run(opts *Options) error {
 		tempFolder,
 		redirectRevisions,
 		opts.Debug,
+		filterOptions,
 	)
 	if err != nil {
 		log.Error().Msgf("❌ Failed to generate apps from ApplicationSets")
 		return err
 	}
+
+	// Check for duplicates again
+	baseApps, targetApps = duplicates.RemoveDuplicates(baseApps, targetApps)
+
+	// enure unique ids
+	baseApps = argoapplication.UniqueIds(baseApps, baseBranch)
+	targetApps = argoapplication.UniqueIds(targetApps, targetBranch)
 
 	if err := utils.CreateFolder(opts.OutputFolder, true); err != nil {
 		log.Error().Msgf("❌ Failed to create output folder: %s", opts.OutputFolder)
